[ANDROSDK-1468] Ktlint, PMD
