+++ core/src/main/java/org/hisp/dhis/android/core/event/internal/EventPostParentCall.kt
@@ -29,11 +29,11 @@ package org.hisp.dhis.android.core.event.internal
 
 import dagger.Reusable
 import io.reactivex.Observable
+import javax.inject.Inject
 import org.hisp.dhis.android.core.arch.call.D2Progress
 import org.hisp.dhis.android.core.event.Event
 import org.hisp.dhis.android.core.tracker.TrackerPostParentCallHelper
 import org.hisp.dhis.android.core.tracker.importer.internal.TrackerImporterPostCall
-import javax.inject.Inject
 
 @Reusable
 internal class EventPostParentCall @Inject internal constructor(
